Block user access after a number of failed attempts
The goal here is to increase the level of security by locking a user
after three failed login attempts. The user will be automatically
unlocked after one hour. The low number (3) of attempts comes from the
fact that this is a single admin system. That should prevent any brute
force attacks on the admin password. All the magic comes from Devise
Lockable module.