Use article slug as Disqus identifier
Using the article ID as Disqus ID cause certain issues with development
and production environments as test articles happen to have the same
identificators as articles in production. Since the public ID of the
article on the web is its friendly URL (slug), it should be also used by
Disqus to identify articles.