Allow article date to be set through the UI
Using the created_at as the article display date causes a number of
issues. For example, I may create a work-in-progress article and it may
stay in editing mode for weeks before it gets published. I would like to
be able to set the article's date manually through the CMS UI. To allow
this behavior I am adding a new date attribute to the Article model.