Add pagination to blog posts
As the blog post numbers are going to grow I need to show them in pages
of smaller numbers. I have chosen 5 articles per page as being a
subjectively sensible number that looks good given the current blog
design. I've decided to use the kaminari gem for pagination due to a
number of reasons, but mainly for being clean and easy to use.