Bring back the Draper gem
I ditched Draper when upgrading to Rails 5 as it was not compatible.
Now with version 3 all seems to be working fine with Rails 5
so I ditch my custom decorator in favor of this gem - no need to solve a
problem that has already been solved by the community, right?

The ArticleDecorator and PaginatingDecorator are mostly needed to have
the Kaminari pagination working with Draper.